requirements.txt
fastapi==0.128.0
sqlalchemy==2.0.46
psycopg2==2.9.11
python-dotenv==1.2.1
pydantic==2.12.5
pytest==9.0.2
uvicorn==0.40.0